now the menu is actually usable.
I've changed "resolution" to "window size" as changing the resolution in a pixel art game serves no benefit.

ever so slightly changed the width and height of some of the maps to fit the the 16 bit tiles. excluding the puzzletestrm map because that fits the the 4/3 /3 display and view port.

made the scrolling inside of the settings menu nicer

the element_options input code's spaghetti math has been fixed.

the element options now returns either false or a function that executes after the rest of the ui has been drawn to avoid overlaps.

deleted the old obj_inventory and get rid of some of the _v2 tags.

im just gona leave it at that for now.
might make is look a bit more jazzy and make the sliding movement smoother latter.